requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
httpx==0.26.0
groq==0.4.2
streamlit==1.31.0
plotly==5.18.0
matplotlib==3.8.2
pandas==2.2.0
numpy==1.26.3
python-multipart==0.0.6
scipy==1.11.4

